Fixed LinkCheck, and refactored Code
